[move/#1] util 패키지 수정
