+++ core/util/src/main/java/com/napzak/market/util/android/NoRippleInteractionSource.kt
@@ -1,4 +1,4 @@
-package com.napzak.market.util
+package com.napzak.market.util.android
 
 import androidx.compose.foundation.interaction.Interaction
 import androidx.compose.foundation.interaction.MutableInteractionSource
