+++ core/util/src/main/java/com/napzak/market/util/android/NumericVisualTransformation.kt
@@ -1,4 +1,4 @@
-package com.napzak.market.util
+package com.napzak.market.util.android
 
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.input.OffsetMapping
